Modified NumericalInputRow so that it uses Integers

+++ src/frontend/wizards/wizard_2_0/wizard_pages/TerrainMovePointPage.java
@@ -42,7 +42,7 @@ public class TerrainMovePointPage extends WizardPage {
 		canNextWritable().setValue(true);
 	}
 
-	public Map<Terrain, Double> getTerrainMovePoints() {
+	public Map<Terrain, Integer> getTerrainMovePoints() {
 		return rowToTerrain.keySet().stream()
 				.collect(Collectors.toMap(row -> rowToTerrain.get(row), row -> row.getValue()));
 	}
